Add files via upload
The "synthetic-data.xlsx" is the file of  synthetic datasets, which were generated by Monte Carlo simulation method. x1 and x2 are the dependent or latent variables that range from -10 to 10. y1 and y2 are the independent or observed variables with binary values equal to 0 or 1 representing the absence or presence accordingly.  The border between y1=0 and y1=1 points is set as a straight line (i.e. x1-x2-5/2=0) denoting a linearly separable case , whereas this border between y2=0 and y2=1 points is defined as a parabolic curve (i.e. 4*x1^2/15-x2-15/2=0) standing for a nonlinearly separable case.
The code of "GEP-LR-synthetic-data-modeling.py" is used for synthetic data modeling.
The code of "GEP-LR-MPM.py" is applied for mineral prospectivity mapping.